COMMERCE-6609 dispatch-talend-web - move expando setup to proper class + make sure each company is updated - we are decoupling MVC layer from one-time low-level initialization
